IG-2552 Update double dropdown for toggle and selects to work and to save values on select/remove
